Refactored Shortcuthandler and ResizeHandler into their own classes.
